seperate form concerns and clean models
