+++ td_biblio/forms/__init__.py
@@ -1,0 +1,7 @@
+# coding: utf-8
+__author__ = 'Alison Mukoma <mukomalison@gmail.com>'
+__license__ = 'MIT'
+__copyright__ = 'tailordev.fr'
+
+from .author_duplicates import * # noqa
+from .entry_batch_import import * # noqa
