+++ td_biblio/models/abstract_entity.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+from django.db import models
+from django.utils.translation import ugettext_lazy as _
+
 class AbstractEntity(models.Model):
     """Simple abstract entity"""
