+++ td_biblio/models/entry.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+from django.db import models
+from django.utils.translation import ugettext_lazy as _
+
+
 class Entry(models.Model):
     """The core model for references
