+++ td_biblio/utils/pmid_validator.py
@@ -1,0 +1,6 @@
+from django.core.validators import RegexValidator
+
+PMID_REGEX = r"^-?\d+\Z"
+pmid_validator = RegexValidator(
+    PMID_REGEX, _("One (or more) PMID is not valid"), "invalid"
+)
